+++ internal/pkg/presentation/api_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/diwise/api-opendata/internal/pkg/application/services/beaches"
 	"github.com/diwise/api-opendata/internal/pkg/application/services/citywork"
-	"github.com/diwise/api-opendata/internal/pkg/application/services/roadaccidents"
 	"github.com/diwise/api-opendata/internal/pkg/application/services/waterquality"
 	"github.com/diwise/api-opendata/internal/pkg/presentation/handlers"
 	"github.com/rs/zerolog"
@@ -19,23 +18,6 @@ import (
 
 func TestMain(m *testing.M) {
 	os.Exit(m.Run())
-}
-
-func TestGetRoadAccidents(t *testing.T) {
-	is := is.New(t)
-
-	w := httptest.NewRecorder()
-	req, _ := http.NewRequest(http.MethodGet, "/api/roadaccidents", nil)
-	req.Header.Add("Accept", "application/json")
-
-	roadAccidentSvc := &roadaccidents.RoadAccidentServiceMock{
-		GetAllFunc: func() []byte { return nil },
-	}
-
-	handlers.NewRetrieveRoadAccidentsHandler(zerolog.Logger{}, roadAccidentSvc).ServeHTTP(w, req)
-
-	is.Equal(w.Code, http.StatusOK)                 // Request failed, status code not OK
-	is.Equal(len(roadAccidentSvc.GetAllCalls()), 1) // should have been called once
 }
 
 func TestGetCitywork(t *testing.T) {
